project/requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.3
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ollama==0.3.3
psycopg2-binary==2.9.10
sniffio==1.3.1
sqlparse==0.5.2

langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.22.0
protobuf~=3.19.0
langchain_experimental
mysql-connector-python
pymysql
sentence-transformers
chromadb
